refactor: comprehensive refactoring based on single responsibility principle
- Extract data layer into separate files with TypeScript interfaces
- Create custom hooks for logic separation (useMobileMenu)
- Break down large components into focused, single-purpose components
- Implement layout components for reusability (Header, Footer, MobileMenu)
- Create section components for better organization
- Separate concerns: data, types, hooks, layout, and presentation
- Reduce home.tsx from 564 lines to 52 lines (91% reduction)
- Improve maintainability and code reusability across all routes
