+++ app/hooks/useMobileMenu.ts
@@ -1,0 +1,28 @@
+import { useState } from "react";
+
+export function useMobileMenu() {
+  const [isMobileMenuOpen, setIsMobileMenuOpen] = useState(false);
+
+  const toggleMobileMenu = () => {
+    setIsMobileMenuOpen(!isMobileMenuOpen);
+  };
+
+  const closeMobileMenu = () => {
+    setIsMobileMenuOpen(false);
+  };
+
+  const scrollToSection = (sectionId: string) => {
+    closeMobileMenu();
+    const element = document.getElementById(sectionId);
+    if (element) {
+      element.scrollIntoView({ behavior: "smooth" });
+    }
+  };
+
+  return {
+    isMobileMenuOpen,
+    toggleMobileMenu,
+    closeMobileMenu,
+    scrollToSection,
+  };
+}
